examples/h5/nisar: rename {products} to {schema} (these are schemas, not on-disk accessors)
